leetcode/celery_flaskazure/requirements.txt
iniconfig==2.0.0
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
packaging==23.2
parse==1.20.0
parse-type==0.6.2
pluggy==1.3.0
pyaml==23.9.7
pytest==7.4.3
pytest-bdd==7.0.0
pytest-html==4.1.1
pytest-metadata==3.0.0
PyYAML==6.0.1
six==1.16.0
typing_extensions==4.8.0
redis
celery